webserver/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aniso8601==9.0.1
    # via flask-restful
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.8
    # via requests
click==8.0.3
    # via flask
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-restful
flask-restful==0.3.9
    # via -r requirements.in
idna==3.3
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.0.1
    # via jinja2
python-dotenv==0.19.2
    # via -r requirements.in
pytz==2021.3
    # via flask-restful
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via flask-restful
urllib3==1.26.7
    # via requests
werkzeug==2.0.2
    # via flask
